+++ package.json
@@ -9,6 +9,7 @@
     "dev:backend": "nodemon backend/server.js",
     "dev:public": "vite --host 0.0.0.0",
     "build": "tsc && vite build",
+    "build:simple": "vite build --mode development",
     "preview": "vite preview",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "start": "node backend/server.js"
