+++ vite.config.ts
@@ -4,9 +4,8 @@ import react from '@vitejs/plugin-react'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-  root: 'frontend',
   build: {
-    outDir: '../frontend/dist',
+    outDir: 'frontend/dist',
   },
   server: {
     port: 5173,
